Update memory allocation strategy in llama3.cu
Removed the comment about using calloc instead of malloc at the top of the malloc_run_state function. The comment was readded above the line where we actually use calloc instead of malloc. The comment checking the success of memory allocations was updated to specifically refer to cudaMalloc instead of general malloc.